+++ agents-flex-chat/agents-flex-chat-openai/src/test/java/com/agentsflex/llm/openai/OpenAIChatModelTest.java
@@ -437,7 +437,7 @@ public class OpenAIChatModelTest {
             @Override
             public void onFinalAnswer(String finalAnswer) {
                 System.out.println("onFinalAnswer >>>>>" + finalAnswer);
-                ChatMemory memory = reActAgent.getHistoriesPrompt().getMemory();
+                ChatMemory memory = reActAgent.getMemoryPrompt().getMemory();
                 System.out.println(memory);
             }