+++ agents-flex-core/src/main/java/com/agentsflex/core/message/SystemMessage.java
@@ -15,7 +15,9 @@
  */
 package com.agentsflex.core.message;
 
-public class SystemMessage extends AbstractTextMessage {
+import java.util.HashMap;
+
+public class SystemMessage extends AbstractTextMessage<SystemMessage> {
 
     public SystemMessage() {
     }
@@ -34,5 +36,20 @@ public class SystemMessage extends AbstractTextMessage {
             "content='" + content + '\'' +
             ", metadataMap=" + metadataMap +
             '}';
+    }
+
+    /**
+     * 创建并返回当前对象的副本。
+     *
+     * @return 一个新的、内容相同但内存独立的对象
+     */
+    @Override
+    public SystemMessage copy() {
+        SystemMessage copy = new SystemMessage();
+        copy.content = this.content;
+        if (this.metadataMap != null) {
+            copy.metadataMap = new HashMap<>(this.metadataMap);
+        }
+        return copy;
     }
 }