+++ agents-flex-core/src/main/java/com/agentsflex/core/message/ToolMessage.java
@@ -15,7 +15,9 @@
  */
 package com.agentsflex.core.message;
 
-public class ToolMessage extends AbstractTextMessage {
+import java.util.HashMap;
+
+public class ToolMessage extends AbstractTextMessage<ToolMessage> {
 
     private String toolCallId;
 
@@ -35,5 +37,21 @@ public class ToolMessage extends AbstractTextMessage {
             ", content='" + content + '\'' +
             ", metadataMap=" + metadataMap +
             '}';
+    }
+
+    /**
+     * 创建并返回当前对象的副本。
+     *
+     * @return 一个新的、内容相同但内存独立的对象
+     */
+    @Override
+    public ToolMessage copy() {
+        ToolMessage copy = new ToolMessage();
+        copy.content = this.content;
+        copy.toolCallId = this.toolCallId;
+        if (this.metadataMap != null) {
+            copy.metadataMap = new HashMap<>(this.metadataMap);
+        }
+        return copy;
     }
 }